removed the valueSetFilter in Salesforce.
- Moved exported utils from the removed filter to the utils.ts module.
